finition back, admin, location

+++ application/views/profile.php
@@ -1,3 +1,5 @@
+<?php var_dump($_SESSION); ?>
+
 <div class="container border">
     <div class="row">
         <div class="col-6 border">
@@ -15,7 +17,7 @@
                         <b>Ville</b> : <?= $profile->u_city ?><br>
                         <b>Numéro de téléphone</b> : <?= $profile->u_phone ?><br>
                         <b>Adresse mail</b> : <?= $profile->u_mail ?><br>
-                        <a type="button" href="form/update<?= "/".$profile->u_id ?>" class="btn btn-light">Modifier</a>
+                        <a type="button" href="form/update" class="btn btn-light">Modifier</a> 
                     </div>
                 <?php
                 }
@@ -33,7 +35,15 @@
                         <b>Véhicule</b> : <?= $location->c_model ?><br> 
                         <b>Date de début</b> : <?= $location->l_startdate ?><br>
                         <b>Date de fin</b> : <?= $location->l_enddate ?><br>
-                        <a href="deleteLocation<?= "/".$location->l_id ?>">Annuler</a>
+                        <?php
+                        $today = date("Y-m-d");
+                        if($location->l_isValid == 0 || ($location->l_isValid == 1 && $location->l_isReturn == 0 && $location->l_startdate > $today))
+                        {
+                            ?>
+                            <a href="deleteLocation<?= "/".$location->l_id ?>">Annuler</a>
+                            <?php
+                        }
+                        ?>
                     </div>
                 </div>
             <?php
